+++ libft/includes/libft.h
@@ -71,7 +71,7 @@ void			ft_putstr_fd(char *s, int fd);
 void			ft_putendl_fd(char *s, int fd);
 void			ft_putnbr_fd(int n, int fd);
 
-double			ft_abs(double n);
+float			ft_abs(float n);
 int				ft_atoi_base(char *str, char *base);
 int				ft_check_base(char *base);
 char			*ft_convert_base(char *nbr, char *base_from, char *base_to);
